// katana-slice-manager/src/app/models/interfaces/nfvo-registration-form.interface.ts
export interface NfvoRegistrationFormModel {
  id: string;
  name: string;
  nfvoip: string;
  nfvousername: string;
  nfvopassword: string;
  tenantname: string;
  type: string;
  version: string;
  description: string;
  configId: string;
  configNfvoUsername: string;
  configNfvoPassword: string;
  configNfvoIp: string;
  configTenantName: string;
}